extract dataclasses for fuels2030
